+++ libs/designer/src/lib/ui/panel/recommendation/recommendationPanelContext.tsx
@@ -1,34 +1,41 @@
-import type { RootState } from '../../../core';
+import type { AppDispatch } from '../../../core';
+import { addOperation } from '../../../core/actions/bjsworkflow/add';
 import { useAllOperations } from '../../../core/queries/browse';
-import { useIsConsumption } from '../../../core/state/designerOptions/designerOptionsSelectors';
-import { useSelectedOperationGroupId } from '../../../core/state/panel/panelSelectors';
-import { selectOperationGroupId } from '../../../core/state/panel/panelSlice';
+import {
+  useIsAddingTrigger,
+  useIsParallelBranch,
+  useRelationshipIds,
+  useSelectedSearchOperationGroupId,
+} from '../../../core/state/panel/panelSelectors';
+import { selectOperationGroupId, selectOperationId } from '../../../core/state/panel/panelSlice';
 import { BrowseView } from './browseView';
 import { OperationGroupDetailView } from './operationGroupDetailView';
 import { SearchView } from './searchView';
-import type { CommonPanelProps } from '@microsoft/designer-ui';
+import { Link, Icon } from '@fluentui/react';
 import { RecommendationPanel, OperationSearchHeader } from '@microsoft/designer-ui';
+import type { CommonPanelProps } from '@microsoft/designer-ui';
 import type { DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/utils-logic-apps';
-import { areApiIdsEqual } from '@microsoft/utils-logic-apps';
+import { guid, areApiIdsEqual } from '@microsoft/utils-logic-apps';
 import { useCallback, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useIntl } from 'react-intl';
+import { useDispatch } from 'react-redux';
 
 export const RecommendationPanelContext = (props: CommonPanelProps) => {
-  const dispatch = useDispatch();
-  const isNodeTrigger = useSelector((state: RootState) => state.panel.addingTrigger);
+  const dispatch = useDispatch<AppDispatch>();
+  const isTrigger = useIsAddingTrigger();
   const [searchTerm, setSearchTerm] = useState('');
   const [filters, setFilters] = useState<Record<string, string>>({
-    actionType: isNodeTrigger ? 'triggers' : 'actions',
+    actionType: isTrigger ? 'triggers' : 'actions',
   });
   const [allOperationsForGroup, setAllOperationsForGroup] = useState<DiscoveryOperation<DiscoveryResultTypes>[]>([]);
 
   const [isGrouped, setIsGrouped] = useState(true);
 
-  const selectedOperationGroupId: string = useSelectedOperationGroupId();
+  const [isSelectingAzureResource, setIsSelectingAzureResource] = useState(false);
+
+  const selectedOperationGroupId = useSelectedSearchOperationGroupId();
 
   const allOperations = useAllOperations();
-
-  const isConsumption = useIsConsumption();
 
   useEffect(() => {
     if (allOperations.data && selectedOperationGroupId) {
@@ -47,37 +54,68 @@ export const RecommendationPanelContext = (props: CommonPanelProps) => {
   }, [dispatch, props]);
 
   const navigateBack = useCallback(() => {
-    if (selectedOperationGroupId) dispatch(selectOperationGroupId(''));
-    else if (searchTerm) setSearchTerm('');
-  }, [dispatch, searchTerm, selectedOperationGroupId]);
+    dispatch(selectOperationGroupId(''));
+    dispatch(selectOperationId(''));
+    setIsSelectingAzureResource(false);
+  }, [dispatch]);
+
+  const relationshipIds = useRelationshipIds();
+  const isParallelBranch = useIsParallelBranch();
+
+  const onOperationClick = useCallback(
+    (id: string) => {
+      const operation = (allOperations.data ?? []).find((o: any) => o.id === id);
+      if (!operation) return;
+      dispatch(selectOperationId(operation.id));
+      const newNodeId = (operation?.properties?.summary ?? operation?.name ?? guid()).replaceAll(' ', '_');
+      dispatch(addOperation({ operation, relationshipIds, nodeId: newNodeId, isParallelBranch, isTrigger }));
+    },
+    [allOperations.data, dispatch, isParallelBranch, isTrigger, relationshipIds]
+  );
+
+  const intl = useIntl();
+  const returnToSearchText = intl.formatMessage({
+    defaultMessage: 'Return to search',
+    description: 'Text for the Details page navigation heading',
+  });
 
   return (
     <RecommendationPanel placeholder={''} {...props}>
-      <OperationSearchHeader
-        searchCallback={setSearchTerm}
-        onGroupToggleChange={() => setIsGrouped(!isGrouped)}
-        isGrouped={isGrouped}
-        searchTerm={searchTerm}
-        selectedGroupId={selectedOperationGroupId}
-        onDismiss={onDismiss}
-        navigateBack={navigateBack}
-        filters={filters}
-        setFilters={setFilters}
-        isTriggerNode={isNodeTrigger}
-        isConsumption={isConsumption}
-      />
+      {isSelectingAzureResource || selectedOperationGroupId ? (
+        <div className={'msla-sub-heading-container'}>
+          <Link onClick={navigateBack} style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
+            <Icon iconName="Back" />
+            {returnToSearchText}
+          </Link>
+        </div>
+      ) : null}
       {selectedOperationGroupId ? (
-        <OperationGroupDetailView groupOperations={allOperationsForGroup} filters={filters} />
-      ) : searchTerm ? (
-        <SearchView
-          searchTerm={searchTerm}
-          allOperations={allOperations.data ?? []}
-          groupByConnector={isGrouped}
-          isLoading={allOperations.isLoading}
-          filters={filters}
-        />
+        <OperationGroupDetailView groupOperations={allOperationsForGroup} filters={filters} onOperationClick={onOperationClick} />
       ) : (
-        <BrowseView filters={filters} />
+        <>
+          <OperationSearchHeader
+            searchCallback={setSearchTerm}
+            onGroupToggleChange={() => setIsGrouped(!isGrouped)}
+            isGrouped={isGrouped}
+            searchTerm={searchTerm}
+            onDismiss={onDismiss}
+            filters={filters}
+            setFilters={setFilters}
+            isTriggerNode={isTrigger}
+          />
+          {searchTerm ? (
+            <SearchView
+              searchTerm={searchTerm}
+              allOperations={allOperations.data ?? []}
+              groupByConnector={isGrouped}
+              isLoading={allOperations.isLoading}
+              filters={filters}
+              onOperationClick={onOperationClick}
+            />
+          ) : (
+            <BrowseView filters={filters} />
+          )}
+        </>
       )}
     </RecommendationPanel>
   );